Fix for HARMONY-801 ([classlib][nio] java.nio.ByteBuffer.put(byte[], int off, Integer.MAX_VALUE) throws java.nio.BufferUnderflowException instead of IndexOutOfBoundsException)
git-svn-id: https://svn.apache.org/repos/asf/incubator/harmony/enhanced/classlib/trunk@419923 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/nio/src/main/java/java/nio/ReadWriteCharArrayBuffer.java
@@ -101,7 +101,7 @@ final class ReadWriteCharArrayBuffer extends CharArrayBuffer {
 	}
 
     public CharBuffer put(char[] src, int off, int len) {
-        if (off < 0 || len < 0 || len + off > src.length) {
+        if (off < 0 || len < 0 || (long)len + (long)off > src.length) {
             throw new IndexOutOfBoundsException();
         }
         if (len > remaining()) {
