+++ modules/nio/src/main/java/java/nio/ReadWriteDirectByteBuffer.java
@@ -101,7 +101,7 @@ final class ReadWriteDirectByteBuffer extends DirectByteBuffer {
      * @see java.nio.ByteBuffer#put(byte[], int, int)
      */
     public ByteBuffer put(byte[] src, int off, int len) {
-        if (off < 0 || len < 0 || off + len > src.length) {
+        if (off < 0 || len < 0 || (long)off + (long)len > src.length) {
             throw new IndexOutOfBoundsException();
         }
         if (len > remaining()) {
